Change: Move repl code to compile module
